StargazingAi/backend/requirements.txt
fastapi
uvicorn
skyfield
httpx
python-dotenv
requests